+++ crack-core/src/test/mod.rs
@@ -1,7 +1,3 @@
 pub mod errors;
 pub mod http_utils;
 pub mod structs;
-
-
-
-
