+++ cracktunes/src/lib.rs
@@ -1,6 +1,2 @@
 #[cfg(test)]
 pub mod test;
-
-// #[allow(unstable_features)]
-// #[feature(trace_macros)]
-// trace_macros!(true);
