+++ cracktunes/src/test/mod.rs
@@ -1,7 +1,3 @@
-// pub mod commands;
 pub mod errors;
 pub mod playlist;
 pub mod utils;
-
-// pub use commands::*;
-
